+++ Classes/Geocoding/Google.php
@@ -5,16 +5,26 @@ use TYPO3\CMS\Core\Utility\GeneralUtility;
 /**
  * This class represents a service to look up geo coordinates via Google Maps.
  *
+ * @see https://developers.google.com/maps/documentation/javascript/geocoding?#GeocodingStatusCodes
+ *
  * @author Oliver Klee <typo3-coding@oliverklee.de>
  */
 class Tx_Oelib_Geocoding_Google implements \Tx_Oelib_Interface_GeocodingLookup
 {
     /**
-     * status code for: okay, address was parsed
-     *
      * @var string
      */
     const STATUS_OK = 'OK';
+
+    /**
+     * @var string
+     */
+    const STATUS_ZERO_RESULTS = 'ZERO_RESULTS';
+
+    /**
+     * @var string
+     */
+    const STATUS_INVALID_REQUEST = 'INVALID_REQUEST';
 
     /**
      * the base URL of the Google Maps geo coding service
@@ -130,6 +140,10 @@ class Tx_Oelib_Geocoding_Google implements \Tx_Oelib_Interface_GeocodingLookup
                 $resultParts = json_decode($response, true);
                 $status = $resultParts['status'];
                 $lookupError = $status !== self::STATUS_OK;
+                $addressIsInvalid = \in_array($status, [self::STATUS_ZERO_RESULTS, self::STATUS_INVALID_REQUEST], true);
+                if ($addressIsInvalid) {
+                    break;
+                }
             }
 
             if ($httpError || $lookupError) {